typescript: Migrate shortcuts-section.js to typescript.
